Revert "test: move kratos e2e tests to integration (#3109)"
This reverts commit f6a36f16445c393e5b20306945529a0cf279c26f.
