+++ src/domain/authentication/index.types.d.ts
@@ -23,39 +23,21 @@ type IdentityPhone = IdentityBase & {
   phone: PhoneNumber
   email: undefined
   emailVerified: undefined
-
-  username?: undefined
 }
 
 type IdentityEmail = IdentityBase & {
   phone: undefined
   email: EmailAddress
   emailVerified: boolean
-
-  username?: undefined
 }
 
 type IdentityPhoneEmail = IdentityBase & {
   phone: PhoneNumber
   email: EmailAddress
   emailVerified: boolean
-
-  username?: undefined
 }
 
-type IdentityDeviceAccount = IdentityBase & {
-  username: IdentityUsername
-
-  phone?: undefined
-  email?: undefined
-  emailVerified?: undefined
-}
-
-type AnyIdentity =
-  | IdentityPhone
-  | IdentityEmail
-  | IdentityPhoneEmail
-  | IdentityDeviceAccount
+type AnyIdentity = IdentityPhone | IdentityEmail | IdentityPhoneEmail
 
 type Session = {
   identity: AnyIdentity
