+++ test/e2e/servers/index.types.d.ts
@@ -1,0 +1,1 @@
+type PID = number & { readonly brand: unique symbol }
